+++ catalog/app/components/FileEditor/FileEditor.tsx
@@ -12,6 +12,7 @@ import type { S3HandleBase } from 'utils/s3paths'
 
 import Skeleton from './Skeleton'
 import TextEditor from './TextEditor'
+import QuiltConfigEditor from './QuiltConfigEditor'
 import { detect, loadMode, useWriteData } from './loader'
 import { EditorInputType } from './types'
 
@@ -40,7 +41,7 @@ interface EditorState {
   onCancel: () => void
   onChange: (value: string) => void
   onEdit: () => void
-  onSave: () => Promise<void>
+  onSave: () => Promise<Model.S3File | void>
   type: EditorInputType
   value?: string
 }
@@ -67,8 +68,10 @@ export function useState(handle: S3HandleBase): EditorState {
       setEditing(false)
       setSaving(false)
       redirect(h)
+      return h
     } catch (e) {
-      setError(e instanceof Error ? e : new Error(`${e}`))
+      const err = e instanceof Error ? e : new Error(`${e}`)
+      setError(err)
       setSaving(false)
     }
   }, [redirect, value, writeFile])
@@ -110,10 +113,23 @@ function EditorSuspended({
   onChange,
   type,
 }: EditorProps) {
-  loadMode(type.brace || 'plain_text') // TODO: loaders#typeText.brace
+  if (type.brace !== '__quiltConfig') {
+    loadMode(type.brace || 'plain_text') // TODO: loaders#typeText.brace
+  }
 
   const data = PreviewUtils.useObjectGetter(handle, { noAutoFetch: empty })
-  if (empty) return <TextEditor error={error} type={type} value="" onChange={onChange} />
+  if (empty)
+    return type.brace === '__quiltConfig' ? (
+      <QuiltConfigEditor
+        handle={handle}
+        disabled={disabled}
+        error={error}
+        onChange={onChange}
+        initialValue=""
+      />
+    ) : (
+      <TextEditor error={error} type={type} value="" onChange={onChange} />
+    )
   return data.case({
     _: () => <Skeleton />,
     Err: (
@@ -124,8 +140,19 @@ function EditorSuspended({
         <PreviewDisplay data={AsyncResult.Err(err)} />
       </div>
     ),
-    Ok: (response: $TSFixMe) => {
+    Ok: (response: { Body: Buffer }) => {
       const value = response.Body.toString('utf-8')
+      if (type.brace === '__quiltConfig') {
+        return (
+          <QuiltConfigEditor
+            handle={handle}
+            disabled={disabled}
+            error={error}
+            onChange={onChange}
+            initialValue={value}
+          />
+        )
+      }
       return (
         <TextEditor
           disabled={disabled}