+++ catalog/app/components/FileEditor/loader.ts
@@ -1,6 +1,8 @@
+import type { S3 } from 'aws-sdk'
 import * as R from 'ramda'
 import * as React from 'react'
 
+import * as quiltConfigs from 'constants/quiltConfigs'
 import { detect as isMarkdown } from 'components/Preview/loaders/Markdown'
 import * as PreviewUtils from 'components/Preview/loaders/utils'
 import * as AWS from 'utils/AWS'
@@ -18,6 +20,12 @@ export const loadMode = (mode: Mode) => {
     cache[mode] = 'fullfilled'
   })
   throw cache[mode]
+}
+
+const isQuiltConfig = (path: string) =>
+  quiltConfigs.all.some((quiltConfig) => quiltConfig.includes(path))
+const typeQuiltConfig: EditorInputType = {
+  brace: '__quiltConfig',
 }
 
 const isCsv = PreviewUtils.extIn(['.csv', '.tsv', '.tab'])
@@ -51,6 +59,7 @@ const typeNone: EditorInputType = {
 export const detect: (path: string) => EditorInputType = R.pipe(
   PreviewUtils.stripCompression,
   R.cond([
+    [isQuiltConfig, R.always(typeQuiltConfig)],
     [isCsv, R.always(typeCsv)],
     [isJson, R.always(typeJson)],
     [isMarkdown, R.always(typeMarkdown)],
@@ -66,21 +75,36 @@ export const isSupportedFileType: (path: string) => boolean = R.pipe(
   Boolean,
 )
 
+type AWSError = Error & { code: string }
+
+async function validToWriteFile(s3: S3, bucket: string, key: string, version?: string) {
+  try {
+    const { VersionId } = await s3.headObject({ Bucket: bucket, Key: key }).promise()
+    return VersionId === version
+  } catch (error) {
+    if (error instanceof Error && (error as AWSError).code === 'NotFound') return true
+    throw error
+  }
+}
+
 export function useWriteData({
   bucket,
   key,
+  version,
 }: S3HandleBase): (value: string) => Promise<Model.S3File> {
   const s3 = AWS.S3.use()
   return React.useCallback(
     async (value) => {
-      const { VersionId: version } = await s3
+      const valid = await validToWriteFile(s3, bucket, key, version)
+      if (!valid) throw new Error('Revision is outdated')
+      const { VersionId } = await s3
         .putObject({ Bucket: bucket, Key: key, Body: value })
         .promise()
       const { ContentLength: size } = await s3
-        .headObject({ Bucket: bucket, Key: key, VersionId: version })
+        .headObject({ Bucket: bucket, Key: key, VersionId })
         .promise()
-      return { bucket, key, size, version }
+      return { bucket, key, size, version: VersionId }
     },
-    [bucket, key, s3],
+    [bucket, key, s3, version],
   )
 }