+++ catalog/app/containers/Bucket/PackageTree/RevisionMenu.tsx
@@ -7,12 +7,14 @@ import Menu from '../Menu'
 
 interface RevisionMenuProps {
   className: string
+  onCreateFile: () => void
   onDelete: () => void
   onDesktop: () => void
 }
 
 export default function RevisionMenu({
   className,
+  onCreateFile,
   onDelete,
   onDesktop,
 }: RevisionMenuProps) {
@@ -21,6 +23,12 @@ export default function RevisionMenu({
 
   const items = React.useMemo(() => {
     const menu = []
+    if (preferences?.ui?.actions?.revisePackage) {
+      menu.push({
+        onClick: onCreateFile,
+        title: 'Create file',
+      })
+    }
     if (preferences?.ui?.actions?.deleteRevision) {
       menu.push({
         onClick: onDelete,
@@ -34,7 +42,7 @@ export default function RevisionMenu({
       })
     }
     return menu
-  }, [desktop, onDelete, onDesktop, preferences])
+  }, [desktop, onCreateFile, onDelete, onDesktop, preferences])
 
   if (!items.length) return null